Fix spikegen.latency to preserve the input tensor's dtype
Problem
-------

`snntorch.spikegen.latency` silently returns `float32` tensors for
`float64` inputs. This occurs on the default path and when using
`normalize`, `linear`, or `interpolate` (#440).

Root cause
----------

`spikegen.py` defines a module-level `dtype = torch.float`, and the
latency path has three hard-coded `float32` assumptions:

1. `latency()` allocates the spike tensor with
   `torch.zeros(..., dtype=dtype)` (module-level constant).
2. `latency_interpolate()` casts the spike times with
   `torch.round(spike_time).float()`.
3. `latency_interpolate()` allocates its output with
   `torch.ones(..., dtype=dtype)` (module-level constant).

The generated spike tensor therefore uses the fixed `float32` dtype
rather than the dtype of the input tensor.

Fix
---

Derive the dtype from the data instead of the module constant:

1. `latency()`: `torch.zeros(..., dtype=data.dtype)`.
2. `latency_interpolate()`: `torch.round(spike_time).to(spike_time.dtype)`
   — rounding already preserves dtype, only the explicit `.float()`
   downcast is removed.
3. `latency_interpolate()`: `torch.ones(..., dtype=spike_time.dtype)`.

The module-level `dtype` constant is left untouched for the other
call sites (e.g. `to_one_hot` operating on integer targets), so this
change is strictly scoped to the latency path. Behavior for
`float32` inputs is unchanged.

Verification
------------

* Reproduction from the issue: a `float64` input now returns
  `float64` for the default, `normalize`, `linear+normalize`, and
  `interpolate+normalize` paths.
* Output values are identical between `float64` and `float32` inputs
  (checked element-wise for both the scatter and interpolate paths).
* Existing `test_spikegen.py` suite still passes.
* New regression tests in
  `tests/test_snntorch/functional/test_quant_dtype.py` cover the
  four latency paths parametrized over float dtypes (landed with
  #439's state_quant dtype tests in the sibling PR).
* Also verified on Ascend NPU (torch 2.14 + torch_npu, 910B): dtype
  is preserved on `npu:0` for `float32`, `float64`, and `bfloat16`
  inputs end-to-end (latency encoding -> Linear -> Leaky neuron).

Fixes #440

+++ snntorch/spikegen.py
@@ -300,7 +300,7 @@ def latency(
 
         spike_data = torch.zeros(
             (tuple([num_steps] + list(spike_time.size()))),
-            dtype=dtype,
+            dtype=data.dtype,
             device=device,
         )
 
@@ -1357,10 +1357,8 @@ def latency_interpolate(spike_time, num_steps, on_target=1, off_target=0):
 
     device = spike_time.device
 
-    spike_time = torch.round(
-        spike_time
-    ).float()  # Needs to be float as 0s and out-of-bounds spikes
-    # are set to 0.5
+    # Needs to be rounded, as 0s and out-of-bounds spikes are set to 0.5
+    spike_time = torch.round(spike_time).to(spike_time.dtype)
 
     spike_time[
         spike_time > num_steps
@@ -1369,7 +1367,7 @@ def latency_interpolate(spike_time, num_steps, on_target=1, off_target=0):
 
     interpolated_targets = torch.ones(
         (tuple([num_steps] + list(spike_time.size()))),
-        dtype=dtype,
+        dtype=spike_time.dtype,
         device=device,
     )
